Adding some cleanup and organization
