+++ packages/terra-data-grid/src/default-components/SectionHeader.jsx
@@ -22,7 +22,7 @@ const SectionHeader = ({ sectionId, text, isCollapsible, isCollapsed, onClick })
     style={{ zIndex: '1001' }}
     className={cx(['section-header', { selectable: isCollapsible }])}
     data-section-id={sectionId}
-    onClick={onClick}
+    onClick={onClick ? ((event) => { onClick(event, sectionId); }) : null}
   >
     {isCollapsible ? (
       <div className={cx('collapsible-icon')}>
