+++ packages/terra-data-grid/src/default-components/SectionHeader.scss
@@ -7,7 +7,6 @@
     display: flex;
     align-items: center;
     padding: 5px;
-    cursor: pointer;
     user-select: none;
 
     .collapsible-icon {
@@ -29,6 +28,7 @@
   }
 
   .section-header.selectable:hover {
+    cursor: pointer;
     background: grey;
   }
 }
